refactor(agent_backends): extract gateway_translate submodule
Move the Anthropic ↔ OpenAI/Codex wire-translation cluster out of the
facade and into gateway_translate.rs:

- GatewayUpstreamError + impls (Display, From<String>, internal,
  from_upstream)
- anthropic_error_type_for, truncate_for_error_message,
  upstream_message_is_permanent_failure
- openai_compatible_bearer_token, openai_compatible_default_base,
  openai_compatible_request_model, is_claude_code_model_alias
- preflight_context_window_check
- proxy_anthropic_messages (LM Studio Anthropic-shape pass-through)
- write_anthropic_error_response
- call_openai_responses, call_codex_responses
- codex_instructions_from_anthropic, codex_input_from_anthropic,
  assistant_input_items_from_anthropic, user_input_items_from_anthropic
- codex_responses_url
- openai_response_from_sse, event_output_index
- transcript_from_anthropic, content_value_text, tools_from_anthropic
- anthropic_message_from_openai

Production callers in mod.rs (still inside the gateway connection
handler) reach the three pub(super) entry points the gateway needs:
proxy_anthropic_messages, write_anthropic_error_response, and
call_openai_responses. write_anthropic_error_response delegates to
mod.rs's write_json_response via super:: until the gateway
extraction in the next commit moves both into a shared submodule.

route_path stays in mod.rs (it's gateway-tier — used by
handle_gateway_connection) and was restored after a sed
overshoot during this extraction.

Tests still live in mod.rs in this commit and reach the pub(super)
items via `use super::gateway_translate::{...};`.

cargo fmt --check, cargo clippy on the three CI crates with
-Dwarnings, and the agent_backends tests (111 passed) all pass.
No public-surface change.
